+++ libnopegl/src/ngpu/vulkan/texture_vk.h
@@ -41,9 +41,9 @@ struct ngpu_texture_vk_wrap_params {
 struct ngpu_texture_vk {
     struct ngpu_texture parent;
     VkFormat format;
-    int bytes_per_pixel;
-    int array_layers;
-    int mipmap_levels;
+    size_t bytes_per_pixel;
+    uint32_t array_layers;
+    uint32_t mipmap_levels;
     VkImage image;
     int wrapped_image;
     VkImageLayout default_image_layout;
